3-print_reversed_list_integer.py

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -9,7 +9,7 @@
 
 void print_python_list_info(PyObject *p)
 {
-	Py_ssize_t size, alloc, i;
+	int size, alloc, i;
 	PyObject *item;
 	
 	size = PyList_Size(p);
